Fixes #30
Currently, the fixed version of `readlinkat` is provided in upstream
[nix](https://github.com/nix-rust/nix) crate only, therefore, we
need to change `nix` to be a git dependency, which is not ideal.
However, a new release should be out soon-ish, hence, I reckon it's
OK to go with it as-is for now, and change to a semver when a new
version gets released.

Changes:
* uses `nix::fcntl::readlinkat` instead of a direct call to
  `libc::readlinkat`
* if read target path is longer than the buffer provided,
  pads the buffer with 0s